refactor: hooks 폴더 index 파일 생성 및 적용
