+++ src/components/learnDetail/memo/MemoDotButton.tsx
@@ -1,6 +1,6 @@
 import { ImageDiv } from '@src/components/common';
 import { MemoConfirmModalKey } from '@src/components/learnDetail/modal/ConfirmModal';
-import useClickOutside from '@src/hooks/useClickOutside';
+import { useClickOutside } from '@src/hooks/common';
 import { MemoState } from '@src/pages/learn/[id]';
 import { MemoData } from '@src/services/api/types/learn-detail';
 import dynamic from 'next/dynamic';
