+++ src/components/learnDetail/memo/MemoForm.tsx
@@ -1,5 +1,5 @@
 import { MemoConfirmModalKey } from '@src/components/learnDetail/modal/ConfirmModal';
-import useClickOutside from '@src/hooks/useClickOutside';
+import { useClickOutside } from '@src/hooks/common';
 import { MemoState } from '@src/pages/learn/[id]';
 import { MemoData } from '@src/services/api/types/learn-detail';
 import { COLOR, FONT_STYLES } from '@src/styles';
