+++ src/hooks/common/index.ts
@@ -1,0 +1,2 @@
+export { default as useBodyScrollLock } from './useBodyScrollLock';
+export { default as useClickOutside } from './useClickOutside';
