+++ src/hooks/landing/index.ts
@@ -1,0 +1,2 @@
+export { default as useIsMobile } from './useIsMobile';
+export { default as useSlideObserver } from './useSlideObserver';
